- plan metric in scientific notation - negative plan metric - tests - removal of npm dependency
