+++ common/package.json
@@ -19,7 +19,6 @@
     "body-parser": "^1.19.0",
     "events": "^3.1.0",
     "express": "^4.17.1",
-    "npm": "6.13.4",
     "nunjucks": "^3.2.0",
     "open": "^7.0.2",
     "parse-xsd-duration": "^0.5.0",
