+++ common/src/PddlPlanParser.ts
@@ -19,7 +19,7 @@ export class PddlPlanParser {
     private readonly plans: Plan[] = [];
     public static readonly planStepPattern = /^\s*((\d+|\d+\.\d+)\s*:)?\s*\((.*)\)\s*(\[\s*(\d+|\d+\.\d+)\s*\])?\s*$/gim;
     private readonly planStatesEvaluatedPattern = /^\s*;?\s*States evaluated[\w ]*:[ ]*(\d*)\s*$/i;
-    private readonly planCostPattern = /[\w ]*(cost|metric)[\D :]*[ ]*(\d*|\d*\.\d*)\s*$/i;
+    private readonly planCostPattern = /[\w ]*(cost|metric)[\D]*:\s*([+-]?\d*(\.\d+)?|[+-]?\d(\.\d+)?[Ee][+-]?\d+)\s*$/i;
 
     private planBuilder: PlanBuilder;
     private endOfBufferToBeParsedNextTime = '';
@@ -206,7 +206,7 @@ export class PlanBuilder {
 
         plan.statesEvaluated = this.statesEvaluated;
         // if cost was not output by the planning engine, use the plan makespan
-        plan.cost = this.cost ? this.cost : this.getMakespan();
+        plan.cost = this.cost ?? this.getMakespan();
 
         return plan;
     }
